Remove training API
 - The training API has been deprecated by onnxruntime itself, and it
   will be much easier to remove it rather than deprecate it.

 - The training API wrapper functions have been replaced by stubs that
   return errors in legacy_types.go.

 - The README mentions the old version required for the training API.

 - The Scalar type has been promoted to onnxruntime_go.go, and a test
   has been added for it.
