+++ setup_env_windows.go
@@ -56,10 +56,6 @@ func platformInitializeEnvironment() error {
 		return fmt.Errorf("Error setting ORT API base: %d", tmp)
 	}
 
-	// we do not initialize the training API on windows (see setup_env.go)
-	// because currently we cannot support the conversion from UTF-8 to wide
-	// character. See https://github.com/yalue/onnxruntime_go/pull/56.
-
 	libraryHandle = handle
 	return nil
 }
